attached_assets/requirements_1754018383023.txt
# the_analyst/requirements.txt

streamlit
langchain
langchain-openai
python-dotenv
pandas
tabulate